+++ tests/testthat/test-wflow_status.R
@@ -46,11 +46,11 @@ rmd_scr <- file.path(s$analysis, "scratch.Rmd")
 file.create(rmd_scr)
 # Publish index.Rmd
 rmd_pub <- file.path(s$analysis, "index.Rmd")
-suppressMessages(wflow_publish(rmd_pub, "Publish the index",
+suppressMessages(wflow_publish(rmd_pub, "Publish the index", view = FALSE,
                                project = site_dir))
 # Publish and then modify about.Rmd to have status Mod for Modified
 rmd_mod <- file.path(s$analysis, "about.Rmd")
-suppressMessages(wflow_publish(rmd_mod, "Publish the about page",
+suppressMessages(wflow_publish(rmd_mod, "Publish the about page", view = FALSE,
                                project = site_dir))
 cat("edit", file = rmd_mod, append = TRUE)
 # license.Rmd still has status Unp for Unpublished